+++ js/main.js
@@ -1,6 +1,8 @@
 import {createPhotoObjects} from './photo.js';
 import {renderPictures} from './picture.js';
+import {initPhotoForm} from './picture.js';
 const pictures = createPhotoObjects();
 renderPictures(pictures);
+initPhotoForm();
